// src/app/Models/SignUpViewModel.ts
export class UserSignupViewModel {
    NameAr: string = "";
    NameEn: string = '';
    UserName: string = '';
    Email: string = "";
    Password: string = "";
  
    Country: string = "";
    City: string = "";
    Street: string = "";
    Phones: string[] = [];
    img: string = "";
}

export class UserMarketerSignupViewModel {
    nameEn: string = "";
    nameAr: string = "";
    email: string = "";
    password: string = "";
   
    country: string = "";
    city: string = "";
    street: string = "";

    phones: string[] = [];

    taxCard: string = " ";
    Img: string = " ";
}